pyproject.toml
[tool.poetry]
name = "iit"
version = "0.1.0"
description = ""
authors = ["tkwa <kwathomas0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.1.2"
transformer-lens = "^1.12.1"
networkx = "^3.2.1"
pytest = "^7.4.4"
pillow = "^10.2.0"
torchvision = "^0.16.2"
matplotlib = "^3.8.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
